refactor(pvp): extract PvpSessionRegistry base + re-parent TurnBased and Rps
The two session registries (`TurnBasedBoardSessionRegistry`,
`RpsSessionRegistry`) shared ~120 lines of identical plumbing: Caffeine
cache wiring, atomic PENDING → LIVE transition, pending-phase timeout,
decline/forfeit/get/consumeForResolution primitives, and the
opponent/initiator/live indexers. Only the LIVE-phase scheduling
differs — turn-based games arm a per-move shot clock; RPS arms a
single simultaneous-pick timer.

Lifts the shared bits to `database.pvp.PvpSessionRegistry<TSession>`:

  - Caffeine-backed ConcurrentMap (via BoundedSessionCache)
  - `register`, `decline`, `consumeForResolution`, `forfeit`, `get`
  - `pendingForOpponent`, `pendingForInitiator`, `liveFor` indexers
  - Protected `transitionPendingToLive` helper for subclass `accept`

`TurnBasedBoardSessionRegistry` becomes a 100-line subclass that adds
the per-move shot clock (`armMoveClock` / `cancelMoveClock`) and
overrides `consumeForResolution` to cancel the clock. The intermediate
`Session` subclass adds `moveNumber`.

`RpsSessionRegistry` becomes a 100-line subclass that overrides
`accept` to schedule its single pick-phase timer and keeps
`recordPick`. The Session subclass adds the `picks` map and
`bothPicked` flag.

Net: -254 lines of duplication. Per-game live-TTL defaults stay
distinct (RPS 6m backstop vs base 12m; RPS pickTtl 2m vs TurnBased
moveTtl 1m) per the consolidating-refactor guardrails.

Mechanical follow-up: 13 call sites that referenced
`<Registry>.Session.State.LIVE` now use `PvpSessionRegistry.Session.State.LIVE`
because Kotlin nested enums don't inherit through subclass names. One
test stub in PvpControllerTest standardised on the 6-arg `any()`
matcher (matching the convention in RpsCommandTest) — the generic
erasure changed the bridge method shape just enough to break MockK's
auto-default matching.
