+++ database/src/main/kotlin/database/connect4/Connect4SessionRegistry.kt
@@ -3,6 +3,7 @@ package database.connect4
 import common.connect4.Connect4Engine
 import database.boardgame.TurnBasedBoardSessionRegistry
 import database.configuration.RegistryScheduler
+import database.pvp.PvpSessionRegistry
 import org.springframework.stereotype.Component
 import java.time.Duration
 import java.time.Instant
@@ -93,7 +94,7 @@ class Connect4SessionRegistry(
     ): Connect4Engine.MoveResult? {
         val session = get(id) ?: return null
         return synchronized(session) {
-            if (session.state != TurnBasedBoardSessionRegistry.Session.State.LIVE) return@synchronized null
+            if (session.state != PvpSessionRegistry.Session.State.LIVE) return@synchronized null
             val mark = session.markFor(discordId) ?: return@synchronized null
             if (mark != session.currentTurn) return@synchronized null
             val result = Connect4Engine.applyMove(session.board, column, mark)
